# specs/001-wiki-editor-backend/contracts/public-api.openapi.yaml
openapi: 3.1.0
info:
  title: Wiki Editor Backend Public API
  version: 0.2.0
  description: Public REST contracts exposed through the API Gateway.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /pages/{pageId}/archive:
    post:
      tags: [Pages]
      summary: Archive page
      operationId: archivePage
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/PageId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArchivePageRequest' }
      responses:
        '200':
          description: Page archived
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageArchiveResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /pages/{pageId}/draft:
    patch:
      tags: [Pages]
      summary: Autosave full draft snapshot
      operationId: autosaveDraft
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/PageId'
        - name: Idempotency-Key
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AutosaveDraftRequest' }
      responses:
        '200':
          description: Draft saved
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftSaveResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Rebase required
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RebaseRequiredResponse' }
  /pages/{pageId}/draft/recover:
    get:
      tags: [Pages]
      summary: Recover latest accepted draft head
      operationId: recoverDraft
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          description: Recovery payload
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftRecoveryResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /pages/{pageId}/backlinks:
    get:
      tags: [Pages]
      summary: Get backlinks and related pages
      operationId: getBacklinks
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          description: Backlinks
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BacklinksResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /editor/metadata:
    get:
      tags: [Editor]
      summary: Get editor catalog, slash-menu, hotkeys, and capabilities
      operationId: getEditorMetadata
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - name: workspace_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: page_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Editor metadata
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EditorMetadataResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /search:
    get:
      tags: [Search]
      summary: Search by title, content, links, and embed references
      operationId: searchPages
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - name: workspace_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, updated_at]
            default: relevance
      responses:
        '200':
          description: Search results
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /files/uploads:
    post:
      tags: [Files]
      summary: Start attachment upload
      operationId: startUpload
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartUploadRequest' }
      responses:
        '201':
          description: Upload session
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSessionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /files/uploads/{uploadId}/complete:
    post:
      tags: [Files]
      summary: Finalize uploaded object
      operationId: completeUpload
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/UploadId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteUploadRequest' }
      responses:
        '200':
          description: File metadata
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /files/{fileId}:
    get:
      tags: [Files]
      summary: Retrieve attachment metadata and download URL
      operationId: getFile
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File metadata
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Files]
      summary: Soft delete attachment metadata
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Deleted file metadata
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestId:
      description: Correlation and request tracking identifier
      schema:
        type: string
  parameters:
    RequestIdHeader:
      in: header
      name: X-Request-Id
      required: false
      schema:
        type: string
    PageId:
      in: path
      name: pageId
      required: true
      schema:
        type: string
        format: uuid
...
    UploadId:
      in: path
      name: uploadId
      required: true
      schema:
        type: string
        format: uuid
    FileId:
      in: path
      name: fileId
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Forbidden:
      description: Authorization failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: State conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
        request_id: { type: string }
...
    AutosaveDraftRequest:
      type: object
      required: [base_revision_no, document]
      properties:
        base_revision_no: { type: integer, format: int64 }
        document: { $ref: '#/components/schemas/PageDocument' }
        client_state:
          type: object
          additionalProperties: true
    ArchivePageRequest:
      type: object
      required: [base_revision_no]
      properties:
        base_revision_no: { type: integer, format: int64 }
    DraftSaveResponse:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        accepted_revision_no: { type: integer, format: int64 }
        accepted_revision_id: { type: string, format: uuid }
        status: { type: string, enum: [draft_saved] }
        document: { $ref: '#/components/schemas/PageDocument' }
        embedded_tables:
          type: array
          items: { $ref: '#/components/schemas/TableEmbedDescriptor' }
    DraftRecoveryResponse:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        accepted_revision_no: { type: integer, format: int64 }
        accepted_revision_id: { type: string, format: uuid }
        document: { $ref: '#/components/schemas/PageDocument' }
    RebaseRequiredResponse:
      type: object
      properties:
        reason:
          type: string
          enum: [stale_revision, stale_patch, session_out_of_date]
        latest_revision_no: { type: integer, format: int64 }
        latest_revision_id: { type: string, format: uuid }
        server_document: { $ref: '#/components/schemas/PageDocument' }
        conflicting_patch_id: { type: string }
...
    PageArchiveResponse:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        status: { type: string, enum: [archived] }
        current_draft_revision_no: { type: integer, format: int64 }
        current_draft_revision_id: { type: string, format: uuid }
        archived_at: { type: string, format: date-time }
...
    BacklinksResponse:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        backlinks:
          type: array
          items: { $ref: '#/components/schemas/PageReference' }
        related_pages:
          type: array
          items: { $ref: '#/components/schemas/PageReference' }
    PageReference:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        title: { type: string }
        relation: { type: string, enum: [backlink, related] }
        matched_block_id: { type: string }
    EditorMetadataResponse:
      type: object
      properties:
        block_catalog:
          type: array
          items: { $ref: '#/components/schemas/EditorBlockCatalogEntry' }
        slash_menu_items:
          type: array
          items: { $ref: '#/components/schemas/SlashMenuItem' }
        hotkeys:
          type: array
          items: { $ref: '#/components/schemas/HotkeyDefinition' }
        capabilities:
          type: object
          additionalProperties:
            type: boolean
        embed_catalog:
          type: array
          items: { $ref: '#/components/schemas/EmbedCatalogEntry' }
...
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/SearchResult' }
    SearchResult:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        title: { type: string }
        match_type: { type: string, enum: [title, content, link, embed_reference] }
        snippet: { type: string }
        updated_at: { type: string, format: date-time }
    StartUploadRequest:
      type: object
      properties:
        workspace_id: { type: string, format: uuid }
        page_id: { type: string, format: uuid }
        filename: { type: string }
        content_type: { type: string }
        size_bytes: { type: integer, format: int64 }
        checksum: { type: string }
    UploadSessionResponse:
      type: object
      properties:
        upload_id: { type: string, format: uuid }
        file_id: { type: string, format: uuid }
        object_key: { type: string }
        upload_url: { type: string, format: uri }
        expires_at: { type: string, format: date-time }
        headers:
          type: object
          additionalProperties:
            type: string
    CompleteUploadRequest:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        checksum: { type: string }
    FileObjectResponse:
      type: object
      properties:
        file_id: { type: string, format: uuid }
        object_key: { type: string }
        filename: { type: string }
        content_type: { type: string }
        size_bytes: { type: integer, format: int64 }
        status: { type: string, enum: [uploading, ready, deleted] }
        download_url:
          type: string
          format: uri
          nullable: true
    PageDocument:
      type: object
      required: [blocks]
      properties:
        blocks:
          type: array
          items: { $ref: '#/components/schemas/DocumentBlock' }
        metadata:
          type: object
          additionalProperties: true
    DocumentBlock:
      type: object
      properties:
        id: { type: string }
        type:
          type: string
          enum: [paragraph, heading, checklist, quote, code, table_embed, page_link, image, file]
        text: { type: string }
        attrs:
          type: object
          additionalProperties: true
        embed: { $ref: '#/components/schemas/TableEmbedDescriptor' }
        link: { $ref: '#/components/schemas/PageLinkReference' }
        attachment: { $ref: '#/components/schemas/AttachmentReference' }
    PageLinkReference:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        title: { type: string }
        link_kind: { type: string, enum: [page_ref] }
    AttachmentReference:
      type: object
      properties:
        file_id: { type: string, format: uuid }
        block_id: { type: string }
        filename: { type: string }
    TableEmbedDescriptor:
      type: object
      properties:
        mws_table_id: { type: string }
        title: { type: string }
        display_config:
          type: object
          additionalProperties: true
        preview_state: { type: string, enum: [ready, cached, degraded] }
        schema:
          type: object
          additionalProperties: true
        preview_rows:
          type: array
          items:
            type: object
            additionalProperties: true
    SlashMenuItem:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        category: { type: string }
        block_type: { type: string }
        supports_hotkey: { type: boolean }
    EditorBlockCatalogEntry:
      type: object
      properties:
        type: { type: string }
        label: { type: string }
        supports_children: { type: boolean }
        supports_collaboration: { type: boolean }
    HotkeyDefinition:
      type: object
      properties:
        key: { type: string }
        action: { type: string }
        description: { type: string }
    EmbedCatalogEntry:
      type: object
      properties:
        source: { type: string }
        label: { type: string }
        allowed: { type: boolean }